requirements.txt
fastAPI
uvicorn
python-dotenv
azure-data-tables